RESEARCH_PACK_TEMPLATE/requirements.txt
# LUKHAS Innovation Research Package Requirements
# Python 3.8+

# Core dependencies
numpy>=1.21.0
pandas>=1.3.0
pyyaml>=5.4.1
python-dotenv>=0.19.0

# Testing
pytest>=7.0.0
pytest-asyncio>=0.18.0
pytest-cov>=3.0.0

# API and async
aiohttp>=3.8.0
httpx>=0.23.0
fastapi>=0.95.0
uvicorn>=0.20.0

# Data processing
jsonschema>=4.0.0
pydantic>=1.10.0

# Visualization (optional)
matplotlib>=3.5.0
seaborn>=0.11.0

# Documentation
markdown>=3.3.0

# Safety & monitoring
structlog>=21.5.0

# Development (optional)
black>=22.0.0
flake8>=4.0.0
mypy>=0.950